+++ src/modules/0.0.19/iasql_functions/entity/index.ts
@@ -1,6 +1,5 @@
 import { CreateDateColumn, Column, Entity, PrimaryColumn } from 'typeorm';
 
-// TODO: DROP
 export enum IasqlOperationType {
   APPLY = 'APPLY',
   SYNC = 'SYNC',
@@ -12,7 +11,6 @@ export enum IasqlOperationType {
   UPGRADE = 'UPGRADE',
 }
 
-// TODO: DROP
 @Entity()
 export class IasqlOperation {
   @PrimaryColumn({
@@ -36,49 +34,6 @@ export class IasqlOperation {
     enum: IasqlOperationType,
   })
   optype: IasqlOperationType;
-
-  @Column({
-    type: 'text',
-    array: true,
-  })
-  params: string[];
-
-  @Column({
-    type: 'text',
-    nullable: true,
-  })
-  output: string;
-
-  @Column({
-    type: 'text',
-    nullable: true,
-  })
-  err: string;
-}
-
-@Entity()
-export class IasqlRpc {
-  @PrimaryColumn({
-    type: 'uuid',
-  })
-  opid: string;
-
-  @CreateDateColumn({
-    type: 'timestamptz',
-  })
-  startDate: Date;
-
-  @Column({
-    type: 'timestamptz',
-    nullable: true,
-  })
-  endDate: Date;
-
-  @Column({ type: 'text' })
-  moduleName: string;
-
-  @Column({ type: 'text' })
-  methodName: string;
 
   @Column({
     type: 'text',
