+++ src/modules/0.0.20/aws_memory_db/sql/before_uninstall.sql
@@ -1,0 +1,8 @@
+ALTER TABLE subnet_group DROP CONSTRAINT check_subnet_group_subnets;
+DROP FUNCTION "check_subnet_group_subnets";
+
+DROP TRIGGER IF EXISTS check_subnets_by_subnet_group ON subnet;
+DROP FUNCTION "check_subnets_by_subnet_group";
+
+ALTER TABLE subnet_group DROP CONSTRAINT IF EXISTS check_subnet_group_subnets_same_vpc;
+DROP FUNCTION "check_subnet_group_subnets_same_vpc";
