+++ src/modules/0.0.20/iasql_functions/index.ts
@@ -1,10 +1,17 @@
 /* THIS MODULE IS A SPECIAL SNOWFLAKE. DON'T LOOK AT IT FOR HOW TO WRITE A REAL MODULE */
 import { ModuleBase } from '../../interfaces';
 import { IasqlOperationType } from './entity';
+import { CustomCallRpc } from './rpcs';
 
-class IasqlFunctions extends ModuleBase {
+export class IasqlFunctions extends ModuleBase {
+  customCall: CustomCallRpc;
   constructor() {
     super();
+    this.sql = {
+      afterInstallSqlPath: 'sql/create_fns.sql',
+      beforeUninstallSqlPath: 'sql/drop_fns.sql',
+    };
+    this.customCall = new CustomCallRpc(this);
     super.init();
   }
   iasqlOperationType = IasqlOperationType;
