+++ src/modules/0.0.20/iasql_functions/rpcs/index.ts
@@ -1,0 +1,16 @@
+import { IasqlFunctions } from '..';
+import { Context, RpcBase } from '../../../interfaces';
+
+export class CustomCallRpc extends RpcBase {
+  module: IasqlFunctions;
+  name = 'custom_call';
+  call = async (ctx: Context, arg1: string, arg2: string) => {
+    return JSON.stringify([{ result: 'I have been called!', arg1, arg2, memo: ctx.memo }]);
+  };
+
+  constructor(module: IasqlFunctions) {
+    super();
+    this.module = module;
+    super.init();
+  }
+}
